Seperate types into types folder
