+++ src/app/Dashboard/page.tsx
@@ -2,31 +2,11 @@ import { url } from "@/globals";
 import { Role } from "../utils/credentials";
 import UserDashboard from "./UserDashboard";
 import dashboardStyles from "./Dashboard.module.css";
+import { Users } from "../Types/solve";
 
 export const dynamic = "force-dynamic";
 export const revalidate = 0; // Ensure no caching between requests
-type AllowedEvents = "3x3" | "4x4" | "3x3oh" | "2x2";
 
-export type UserEvent = {
-  event: AllowedEvents;
-  rounds: number[][];
-};
-
-export type UserComp = {
-  competitionId: string;
-  events: UserEvent[];
-};
-
-export type User = {
-  _id: string;
-  username: string;
-  password: string;
-  role: Role;
-  competitions: UserComp[];
-  group: 1 | 2;
-};
-
-type Users = User[];
 async function getUsers(): Promise<
   { success: false } | { parsed: Users; success: true; status: number }
 > {
